Remove redux (breaks calendar). Move remaining components to bundles.
