+++ app/javascript/App/bundles/CalendarPage/BodySection.js
