+++ app/javascript/App/bundles/CalendarPage/CalendarListEventHeader.js
