+++ app/javascript/App/bundles/CalendarPage/Tips.js
