+++ app/javascript/App/bundles/CoinIndex/ColumnNames.js
