+++ app/javascript/App/bundles/CoinIndex/RedGreenSpan.js
