+++ app/javascript/App/bundles/CoinIndex/index.js
