+++ app/javascript/App/bundles/CoinShow/LinksData.js
