+++ app/javascript/App/bundles/CoinShow/index.js
