+++ app/javascript/App/bundles/common/normalizers/coins.js
