+++ app/javascript/App/bundles/common/normalizers/index.js
@@ -1,0 +1,5 @@
+import coins from './coins'
+
+export default {
+  coins,
+}
