+++ app/javascript/App/clientStartup.js
@@ -5,10 +5,10 @@ import ClientApp from './ClientApp'
 
 // non-react-router components
 import NavUserContainer from './bundles/common/containers/NavUserContainer'
-import CoinShow from './components/CoinShow'
+import CoinShow from './bundles/CoinShow'
 import FlashMessageListContainer from './bundles/common/containers/FlashMessageListContainer'
 import WatchButton from './bundles/common/components/WatchButton'
-import CalendarPage from './components/CalendarPage'
+import CalendarPage from './bundles/CalendarPage'
 
 // Register the components so they can be rendered from Rails
 ReactOnRails.register({
