--- app/javascript/App/containers/app/index.js
@@ -1,14 +1,0 @@
-/*
- * Any selectors or actions defined here are made globally available.
- */
-import { createHOC } from '../../lib/redux'
-import * as actions from './actions'
-import * as selectors from './selectors'
-
-export default createHOC({
-  actions,
-  selectors,
-  onMount({ props }) {
-    props.fetchUser()
-  }
-})
