--- app/javascript/App/containers/app/userReducer.js
@@ -1,14 +1,0 @@
-import { fromJS } from 'immutable'
-
-const initialState = false
-
-export default (state = initialState, action) => {
-  const { type, response } = action
-  switch (type) {
-    case 'SET_USER':
-      if (response) return fromJS(response)
-      return state
-    default:
-      return state
-  }
-}
