--- app/javascript/App/lib/misc.js
@@ -1,7 +1,0 @@
-//TODO: Remove - Redux only
-import * as _ from 'lodash'
-import inflection from 'lodash-inflection'
-_.mixin(inflection)
-
-export const pluralize = (word) => _.pluralize(word)
-export const singularize = (word) => _.singularize(word)
