--- app/javascript/App/lib/redux/index.js
@@ -1,25 +1,0 @@
-import createHOC from './createHOC'
-import createEntityReducer from './createEntityReducer'
-import createEntityActions from './createEntityActions'
-import createEntitySagas from './createEntitySagas'
-import createEntitySelectors from './createEntitySelectors'
-import createFilterReducer from './createFilterReducer'
-import createFilterActions from './createFilterActions'
-import createFilterSagas from './createFilterSagas'
-import createFilterSelectors from './createFilterSelectors'
-import mergeReducers from './mergeReducers'
-import apiSagas from './apiSagas'
-
-export {
-  apiSagas,
-  createHOC,
-  createEntityReducer,
-  createEntityActions,
-  createEntitySagas,
-  createEntitySelectors,
-  createFilterReducer,
-  createFilterActions,
-  createFilterSagas,
-  createFilterSelectors,
-  mergeReducers
-}
