--- app/javascript/App/lib/redux/initialState.js
@@ -1,10 +1,0 @@
-import { fromJS } from 'immutable'
-
-export default fromJS({
-  activeEntity: null,
-  activeFilters: [],
-  entityDetails: {},
-  entityIDs: {},
-  entityList: {},
-  loadingEntities: {}
-})
