--- app/javascript/App/normalizers/index.js
@@ -1,7 +1,0 @@
-import calendarEvents from './calendarEvents'
-import coins from './coins'
-
-export default {
-  calendarEvents,
-  coins,
-}
